grade regulation

+++ app.js
@@ -1,8 +1,10 @@
+
 var express = require("express"),
   multer = require("multer"),
   app = express(),
   bodyParser = require("body-parser"),
   passport = require('passport'),
+  flash = require("connect-flash");
   LocalStrategy = require('passport-local'),
   User = require('./models/user'),
   mongoose = require("mongoose");
@@ -12,6 +14,7 @@ app.use(bodyParser.urlencoded({
   extended: true
 }));
 app.set("view engine", "ejs");
+app.use(flash());
 app.use(express.static(__dirname + "/public"));
 
 app.use(require('express-session')({
@@ -38,7 +41,17 @@ var upload = multer({
 });
 var fs = require('fs');
 
-var resultSchema = new Schema({
+var resultsList = new Schema({
+  listid: String,
+  year: String,
+  Regulation: String,
+  Paper: String,
+  system: String,
+  Date: String
+});
+
+
+var resultSchemaMarks = new Schema({
   hallticket: String,
   subject_code: String,
   subject_name: String,
@@ -50,15 +63,18 @@ var resultSchema = new Schema({
 
 });
 
-var resultsList = new Schema({
-  listid: String,
-  year: String,
-  Regulation: String,
-  Paper: String,
-  Date: String
-});
+
+var resultSchemaGrade = new Schema({
+  hallticket: String,
+  subject_code: String,
+  subject_name: String,
+  grade: String,
+  grade_points: Number
+ });
+
 
 var results_list = mongoose.model('results_List', resultsList);
+
 
 
 
@@ -100,12 +116,26 @@ app.get("/admin/addresults", isLoggedIn, function (req, res) {
 
 
 
-app.post("/admin/addresults", isLoggedIn, upload.array('file'), function (req, res) {
+app.post("/admin/addresults", isLoggedIn, upload.array('file'), function (req, res) 
+{
+
+
+  if(req.body.system === "marks")
+
+  {
+
+
+
+
+
+
+
+
   console.log(req.body);
   console.log(req.files);
 
-  var collName = req.body.Year + "_" + req.body.Regulation + "_" + req.body.Paper + "_" + 'Results' + "_" + req.body.Date;
-  let results_model = mongoose.model(collName, resultSchema);
+  var collName = req.body.Year + "_" + req.body.Regulation + "_" + req.body.Paper + "_" + 'Results' + "_" + req.body.Date  + "_" + req.body.system;
+  let results_model = mongoose.model(collName, resultSchemaMarks);
   var InsertedToList = false;
 
   req.files.forEach(function (file) {
@@ -166,6 +196,74 @@ app.post("/admin/addresults", isLoggedIn, upload.array('file'), function (req, r
   });
   res.send(`<script>window.location.href="/admin"</script>`);
 
+}
+else if(req.body.system === "grade")
+
+  {
+console.log(req.body);
+  console.log(req.files);
+
+  var collName = req.body.Year + "_" + req.body.Regulation + "_" + req.body.Paper + "_" + 'Results' + "_" + req.body.Date + "_" + req.body.system;
+  let results_model = mongoose.model(collName, resultSchemaGrade);
+  var InsertedToList = false;
+
+  req.files.forEach(function (file) {
+
+    var result = excelToJson({
+      sourceFile: file.path,
+      header: {
+        // Is the number of rows that will be skipped and will not be present at our result object. Counting from top to bottom
+        rows: 5 // 2, 3,les);
+      },
+      columnToKey: {
+        A: 'hallticket',
+        B: 'subject_code',
+        C: 'subject_name',
+        D: 'grade',
+        E: 'grade_points',
+      }
+    });
+
+    result.Sheet0.forEach(function (years) {
+      years.year = req.body.Year;
+    });
+
+    console.log(result.Sheet0);
+    results_model.insertMany(result.Sheet0, function (err) {
+      if (err)
+        res.send("Error");
+      else {
+        fs.unlink(file.path, function (error) {
+          if (error) {
+            console.log(error);
+          } else {
+            console.log('file_Deleted');
+
+            if (!InsertedToList) {
+              results_list.create({
+                listid: collName,
+                year: req.body.Year,
+                Regulation: req.body.Regulation,
+                Paper: req.body.Paper,
+                date: req.body.Date
+              }, function (error) {
+                if (error) {
+                  console.log(error);
+
+
+                }
+
+              });
+              InsertedToList = true;
+            }
+          }
+        });
+      }
+    });
+  });
+  res.send(`<script>window.location.href="/admin"</script>`);
+
+}
 });
 
 
@@ -188,8 +286,16 @@ app.get("/admin/editresults", isLoggedIn, function (req, res) {
 
 
 app.get("/admin/editresults/:id", isLoggedIn, function (req, res) {
-  console.log(req.params.id);
-  let results_model = mongoose.model(req.params.id, resultSchema);
+
+
+var id = req.params.id;
+var lastChar = id.substr(id.length - 1);
+
+if(lastChar === "s")
+{
+
+   console.log(req.params.id);
+  let results_model = mongoose.model(req.params.id, resultSchemaMarks);
   results_model.collection.drop();
   let results_list = mongoose.model("results_List", resultsList);
   results_list.deleteOne({
@@ -198,6 +304,28 @@ app.get("/admin/editresults/:id", isLoggedIn, function (req, res) {
     if (err) return handleError(err);
     res.send(`<script>window.location.href="/admin/editresults"</script>`);
   });
+
+
+}
+
+else if(lastChar === "e")
+{
+   console.log(req.params.id);
+  let results_model = mongoose.model(req.params.id, resultSchemaGrade);
+  results_model.collection.drop();
+  let results_list = mongoose.model("results_List", resultsList);
+  results_list.deleteOne({
+    'listid': req.params.id
+  }, function (err) {
+    if (err) return handleError(err);
+    res.send(`<script>window.location.href="/admin/editresults"</script>`);
+  });
+
+
+}
+
+ 
+
 
 
 });
@@ -228,7 +356,14 @@ app.get("/results/:id", function (req, res) {
 });
 
 app.post("/results", function (req, res) {
-  let results_model = mongoose.model(req.body.collection_id, resultSchema);
+
+var id = req.body.collection_id;
+var lastChar = id.substr(id.length - 1);
+
+if(lastChar === "s")
+{
+
+  let results_model = mongoose.model(req.body.collection_id, resultSchemaMarks);
 
 // var res = req.body.hallticket_no.toUpperCase(); ;
   results_model.find({
@@ -244,6 +379,29 @@ app.post("/results", function (req, res) {
       // res.send(member);  
     }
   });
+
+}
+
+else if(lastChar === "e")
+{
+  let results_model = mongoose.model(req.body.collection_id, resultSchemaGrade);
+
+// var res = req.body.hallticket_no.toUpperCase(); ;
+  results_model.find({
+    'hallticket': req.body.hallticket_no.toUpperCase()
+  }, function (err, member) {
+    if (err) {
+      console.log(err);
+    } else {
+      res.render("result", {
+        student: member,
+        id: req.body.collection_id
+      });
+      // res.send(member);  
+    }
+  });
+
+}
 });
 
 
@@ -277,24 +435,24 @@ function isLoggedIn(req, res, next) {
 
 
 
-// app.get('/register', function (req, res) {
-//   res.render('register');
-// });
+app.get('/register', function (req, res) {
+  res.render('register');
+});
 
 
-// app.post('/register', function (req, res) {
-//   var newUser = new User({ username: req.body.username }); // Note password NOT in new User
-//   User.register(newUser, req.body.password, function (err, user) {
-//     if (err) {
-//       console.log(err);
-//       return res.render('register');
-//     } else {
-//       passport.authenticate("local")(req, res, function () {
-//         res.redirect('/admin');
-//       });
-//     }
-//   });
-// });
+app.post('/register', function (req, res) {
+  var newUser = new User({ username: req.body.username }); // Note password NOT in new User
+  User.register(newUser, req.body.password, function (err, user) {
+    if (err) {
+      console.log(err);
+      return res.render('register');
+    } else {
+      passport.authenticate("local")(req, res, function () {
+        res.redirect('/admin');
+      });
+    }
+  });
+});
 
 
 app.listen(3000, process.env.IP, function () {
